docs: add visability none for popit and 1 event in js
