+++ src/index.html
@@ -22,6 +22,23 @@
         </div>
       </div>
     </header>
+    <div class="link" id="links">
+      <div>
+        <h5 class="link_gender">Выберите ваш пол</h5>
+      </div>
+      <div>
+        <h5 class="link_age">Укажите ваш возраст</h5>
+      </div>
+      <div>
+        <h5 class="link_groth">Укажите ваш рост</h5>
+      </div>
+      <div>
+        <h5 class="link_eight">Укажите ваш вес</h5>
+      </div>
+      <div>
+        <h5 class="link_fisical">Выберите физическую активность</h5>
+      </div>
+    </div>
     <div class="container container_genders">
       <h3>Давай выберем пол</h3>
       <div class="genders">
