+++ src/js/index.js
@@ -1,16 +1,5 @@
-function eventBtn() {
-  const popatResult = document.getElementById('result');
-  const btnReult = document.querySelector('div.btn');
-
-  btnReult.addEventListener('click', (event) => {
-    const btnRes = event.target.closest('button.btn_calculation');
-    const btnCler = event.target.closest('button.btn_clear');
-    if (btnRes) {
-      popatResult.classList.add('open');
-    } else if (btnCler) {
-      popatResult.classList.remove('open');
-    }
-  });
-}
+import eventBtn from './eventBtn.js';
+import eventLinks from './eventLinks.js';
 
 eventBtn();
+eventLinks();
