+++ src/scss/eight.scss
@@ -1,38 +1,55 @@
-.eight {
-  display: flex;
-  justify-content: space-around;
-  flex-direction: row;
-  margin: 0 20px 20px;
+.container_eight {
+  position: fixed;
+  top: 25%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin: auto;
+  max-height: 35vh;
+  opacity: 0;
+  visibility: hidden;
 
-  .eight_input {
+  .eight {
     display: flex;
-    align-items: center;
-    input {
-      font-size: $_27px;
-      text-align: center;
-      max-width: 80px;
-      font-family: 'Oswald', sans-serif;
-      border: 3px solid $black;
-      border-radius: 3px;
-      background-color: $blue;
-    }
-    input::-webkit-outer-spin-button,
-    input::-webkit-inner-spin-button {
-      -webkit-appearance: none;
-      margin: 0;
-    }
-  }
-  p {
-    max-width: 60%;
-    margin: auto 0;
+    justify-content: space-around;
+    flex-direction: row;
+    margin: 0 20px 20px;
 
-    span {
-      font-size: $_27px;
-      text-transform: uppercase;
-      color: $blue;
+    .eight_input {
+      display: flex;
+      align-items: center;
+      input {
+        font-size: $_27px;
+        text-align: center;
+        max-width: 80px;
+        font-family: 'Oswald', sans-serif;
+        border: 3px solid $black;
+        border-radius: 3px;
+        background-color: $blue;
+      }
+      input::-webkit-outer-spin-button,
+      input::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+      }
     }
-    .eight_secondSpan {
-      color: $deep_yello;
+    p {
+      max-width: 60%;
+      margin: auto 0;
+
+      span {
+        font-size: $_27px;
+        text-transform: uppercase;
+        color: $blue;
+      }
+      .eight_secondSpan {
+        color: $deep_yello;
+      }
     }
   }
+}
+
+.show {
+  opacity: 1;
+  visibility: visible;
 }
