+++ src/scss/fisical.scss
@@ -1,120 +1,137 @@
-.fisical {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
+.container_fisical {
+  position: fixed;
+  top: 15%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin: auto;
+  max-height: 75vh;
+  opacity: 0;
+  visibility: hidden;
 
-  .minimal {
-    border: 1px solid $dark_blue;
-    position: relative;
-    transition: 0.2s linear;
+  .fisical {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
 
-    label {
-      color: $dark_blue;
-    }
-
-    &:hover {
-      background-color: $dark_blue;
+    .minimal {
+      border: 1px solid $dark_blue;
+      position: relative;
+      transition: 0.2s linear;
 
       label {
-        color: $white;
+        color: $dark_blue;
+      }
+
+      &:hover {
+        background-color: $dark_blue;
+
+        label {
+          color: $white;
+        }
+      }
+    }
+
+    .low {
+      border: 1px solid $blue;
+      position: relative;
+      transition: 0.2s linear;
+
+      label {
+        color: $blue;
+      }
+
+      &:hover {
+        background-color: $blue;
+
+        label {
+          color: $white;
+        }
+      }
+    }
+
+    .average {
+      border: 1px solid $yello;
+      position: relative;
+      transition: 0.2s linear;
+
+      label {
+        color: $yello;
+      }
+
+      &:hover {
+        background-color: $yello;
+
+        label {
+          color: $white;
+        }
+      }
+    }
+
+    .high {
+      border: 1px solid $deep_yello;
+      position: relative;
+      transition: 0.2s linear;
+
+      label {
+        color: $deep_yello;
+      }
+
+      &:hover {
+        background-color: $deep_yello;
+
+        label {
+          color: $white;
+        }
+      }
+    }
+
+    .veryHigh {
+      border: 1px solid $bron;
+      position: relative;
+      transition: 0.2s linear;
+
+      label {
+        color: $bron;
+      }
+
+      &:hover {
+        background-color: $bron;
+
+        label {
+          color: $white;
+        }
+      }
+    }
+
+    .fisical_box {
+      display: inline-block;
+      border-radius: 5px;
+      max-width: 80%;
+      margin: 10px auto;
+      padding: 0px 10px 15px;
+      min-width: 50%;
+
+      input {
+        position: absolute;
+        top: -2px;
+        left: 50%;
+        width: 1px;
+        height: 1px;
+        padding: 0;
+        overflow: hidden;
+        white-space: nowrap;
+        border: 0;
+      }
+      label {
+        text-align: center;
+        font-size: $_27px;
       }
     }
   }
+}
 
-  .low {
-    border: 1px solid $blue;
-    position: relative;
-    transition: 0.2s linear;
-
-    label {
-      color: $blue;
-    }
-
-    &:hover {
-      background-color: $blue;
-
-      label {
-        color: $white;
-      }
-    }
-  }
-
-  .average {
-    border: 1px solid $yello;
-    position: relative;
-    transition: 0.2s linear;
-
-    label {
-      color: $yello;
-    }
-
-    &:hover {
-      background-color: $yello;
-
-      label {
-        color: $white;
-      }
-    }
-  }
-
-  .high {
-    border: 1px solid $deep_yello;
-    position: relative;
-    transition: 0.2s linear;
-
-    label {
-      color: $deep_yello;
-    }
-
-    &:hover {
-      background-color: $deep_yello;
-
-      label {
-        color: $white;
-      }
-    }
-  }
-
-  .veryHigh {
-    border: 1px solid $bron;
-    position: relative;
-    transition: 0.2s linear;
-
-    label {
-      color: $bron;
-    }
-
-    &:hover {
-      background-color: $bron;
-
-      label {
-        color: $white;
-      }
-    }
-  }
-
-  .fisical_box {
-    display: inline-block;
-    border-radius: 5px;
-    max-width: 80%;
-    margin: 10px auto;
-    padding: 0px 10px 15px;
-    min-width: 50%;
-
-    input {
-      position: absolute;
-      top: -2px;
-      left: 50%;
-      width: 1px;
-      height: 1px;
-      padding: 0;
-      overflow: hidden;
-      white-space: nowrap;
-      border: 0;
-    }
-    label {
-      text-align: center;
-      font-size: $_27px;
-    }
-  }
+.show {
+  opacity: 1;
+  visibility: visible;
 }
