+++ src/scss/gender.scss
@@ -1,4 +1,15 @@
 .container_genders {
+  position: fixed;
+  top: 23%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin: auto;
+  max-height: 60vh;
+  opacity: 0;
+  visibility: hidden;
+  transition: 0.2s linear;
+
   h3 {
     margin: 15px 0;
   }
@@ -66,4 +77,9 @@
       }
     }
   }
+}
+
+.show {
+  opacity: 1;
+  visibility: visible;
 }
