+++ src/scss/groth.scss
@@ -1,42 +1,59 @@
-.groth {
-  display: flex;
-  justify-content: space-around;
-  flex-direction: row;
+.container_groth {
+  position: fixed;
+  top: 25%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin: auto;
+  max-height: 35vh;
+  opacity: 0;
+  visibility: hidden;
 
-  .groth_input {
+  .groth {
     display: flex;
-    align-items: center;
-    input {
-      font-size: $_27px;
-      text-align: center;
-      max-width: 80px;
-      font-family: 'Oswald', sans-serif;
-      border: 3px solid $black;
-      border-radius: 3px;
-      background-color: $blue;
+    justify-content: space-around;
+    flex-direction: row;
+
+    .groth_input {
+      display: flex;
+      align-items: center;
+      input {
+        font-size: $_27px;
+        text-align: center;
+        max-width: 80px;
+        font-family: 'Oswald', sans-serif;
+        border: 3px solid $black;
+        border-radius: 3px;
+        background-color: $blue;
+      }
+
+      input::-webkit-outer-spin-button,
+      input::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+      }
+    }
+    p {
+      width: 50%;
+      margin: auto 0;
+
+      span {
+        font-size: $_27px;
+        text-transform: uppercase;
+        color: $blue;
+      }
+      .groth_secondSpan {
+        color: $deep_yello;
+      }
     }
 
-    input::-webkit-outer-spin-button,
-    input::-webkit-inner-spin-button {
-      -webkit-appearance: none;
-      margin: 0;
+    .img_groth {
+      margin-bottom: 20px;
     }
   }
-  p {
-    width: 50%;
-    margin: auto 0;
+}
 
-    span {
-      font-size: $_27px;
-      text-transform: uppercase;
-      color: $blue;
-    }
-    .groth_secondSpan {
-      color: $deep_yello;
-    }
-  }
-
-  .img_groth {
-    margin-bottom: 20px;
-  }
+.show {
+  opacity: 1;
+  visibility: visible;
 }
